Rename models to factories because it makes more sense.
